✨ feat : NoticeList  -> 부모인 Notice로 useEffect 이동
