+++ src/components/Notice/NoticeSearchFilter.jsx
@@ -3,12 +3,14 @@ import Input from '@/components/Input'
 import { useId } from 'react'
 
 function NoticeSearchFilter({ option, text, onChangeOption, onChangeText }) {
-  // option, text Notice 상태를 전달받음
+  //  Notice state, setState (option, text)상태를 전달받음
   const id = useId()
 
   const handleInput = (e) => {
     onChangeText(e.target.value)
   }
+
+  const handleSearchButton = () => {}
 
   return (
     <form>
@@ -28,10 +30,12 @@ function NoticeSearchFilter({ option, text, onChangeOption, onChangeText }) {
         <option value="noticeDescription">내용</option>
         <option value="noticeWriter">작성자</option>
       </select>
-      <Input value={text} onChange={handleInput} placeholder="검색어를 입력하세요" className="bg-white mr-[0.3125rem] border px-jj_15 w-fit"></Input>
-      <Button color="primary" className="px-5 py-[0.625rem] mr-[0.3125rem] mobile:mt-[0.9375rem] mobile:w-full">
+      <Input value={text} type="search" onChange={handleInput} placeholder="검색어를 입력하세요" className="bg-white mr-[0.3125rem] border px-jj_15 w-fit"></Input>
+      {/* input type search value입력시 x 표시 나오고 클릭시 텍스트 삭제  */}
+      <Button onClick={handleSearchButton} color="primary" className="px-5 py-[0.625rem] mr-[0.3125rem] mobile:mt-[0.9375rem] mobile:w-full">
         검색
       </Button>
+
       <Button color="primary" className="px-5  mobile:w-full mobile:my-[0.9375rem]">
         등록
       </Button>
